requirements.txt
pysimplegui
imageio
onnxruntime
opencv-python
numpy
Pillow